+++ R/module_snapshot_manager.R
@@ -183,6 +183,12 @@ srv_snapshot_manager <- function(id, slices_global) {
         modalDialog(
           easyClose = TRUE,
           textInput(ns("snapshot_name"), "Name the snapshot", width = "100%", placeholder = "Meaningful, unique name"),
+          tags$script(
+            shiny::HTML(
+              sprintf("shinyjs.autoFocusModal('%s');", ns("snapshot_name")),
+              sprintf("shinyjs.enterToSubmit('%s', '%s');", ns("snapshot_name"), ns("snapshot_name_accept"))
+            )
+          ),
           footer = shiny::div(
             shiny::tags$button(
               type = "button",