+++ NEWS.md
@@ -1,5 +1,11 @@
 # teal 1.0.0.9002
 
+### New features
+
+* `init` and `srv_teal` have new `reporter` parameter, that allows to pre-define `teal.reporter::Reporter` object to be
+used for storing the content of the report. You can also globally disable reporting by setting `reporter = NULL`
+(and `disable = TRUE` in `ui_teal` for cases when `ui_teal` is used as shiny module).
+
 # teal 1.0.0
 
 ### Breaking changes